apps/worker/requirements.txt
redis
sqlalchemy
psycopg2-binary
python-dotenv
cloudinary
gitpython
requests